requirements.txt
absl-py==0.7.1
astor==0.7.1
audioread==2.1.6
backcall==0.1.0
cycler==0.10.0
decorator==4.4.0
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
ipython==7.4.0
ipython-genutils==0.2.0
jedi==0.13.3
joblib==0.13.2
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
librosa==0.6.3
llvmlite==0.28.0
Markdown==3.1
matplotlib==3.0.3
mock==2.0.0
numba==0.43.1
numpy==1.16.2
pandas==0.24.2
parso==0.4.0
pbr==5.1.3
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkg-resources==0.0.0
prompt-toolkit==2.0.9
protobuf==3.7.1
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2018.9
resampy==0.2.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.15.2
